Added login and messaging with recyclerviews

+++ MessageApp/app/src/main/java/com/example/messageapp/code/database/CRUD.kt
@@ -1,8 +1,12 @@
 package com.example.messageapp.code.database
 
+/**
+ * Crud interface
+ */
+
 interface CRUD <T>{
     fun create(t : T) : Long
-    fun read(id : Int) : T
-    fun update(id : Int, t: T)
-    fun delete(id: Int) : Long
+    fun read(id : Long) : T
+    fun update(id : Long, t: T)
+    fun delete(id: Long) : Long
 }
